1단계 학습 테스트 클래스 분리 및 리팩토링
