+++ src/test/java/study/SetCollectionTest.java
@@ -1,0 +1,57 @@
+package study;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class SetCollectionTest {
+
+    private Set<Integer> numbers;
+
+    @BeforeEach
+    void setUp() {
+        numbers = new HashSet<>();
+        numbers.add(1);
+        numbers.add(1);
+        numbers.add(2);
+        numbers.add(3);
+    }
+
+    @Test
+    @DisplayName("Set 크기 확인")
+    void size() {
+        assertThat(numbers.size()).isEqualTo(3);
+    }
+
+    @Test
+    @DisplayName("Set 데이터 추가 후 크기 확인")
+    void sizeAdd() {
+        numbers.add(4);
+        numbers.add(5);
+
+        assertThat(numbers.size()).isEqualTo(5);
+    }
+
+    @ParameterizedTest
+    @ValueSource(ints = {1, 2, 3})
+    @DisplayName("값이 존재하는지 확인")
+    void contains(int input) {
+        assertThat(numbers.contains(input)).isTrue();
+    }
+
+    @ParameterizedTest
+    @CsvSource(value= {"1:true", "2:true", "3:true", "4:false", "5:false"}, delimiter = ':')
+    @DisplayName("입력 값에 따라 결과 값이 다른 경우")
+    void containsFalse(int input, boolean expected) {
+        Assertions.assertEquals(numbers.contains(input), expected);
+    }
+}
